refactor(cli): cleanup signup command
- [x] Remove signup command from cli
- [x] Move gatekeeper getters to sdk
