+++ packages/cli/ts/index.ts
@@ -852,8 +852,6 @@ export {
   publishBatch,
   proveOnChain,
   setVerifyingKeysCli,
-  signup,
-  isRegisteredUser,
   timeTravel,
   verify,
   joinPoll,
@@ -868,17 +866,14 @@ export type {
   GenLocalStateArgs,
   GenProofsArgs,
   PublishArgs,
-  SignupArgs,
   MergeSignupsArgs,
   VerifyArgs,
   ProveOnChainArgs,
   DeployArgs,
-  IRegisteredUserArgs,
   IGenKeypairArgs,
   IGetPollArgs,
   IGetPollData,
   IPublishBatchArgs,
   IPublishBatchData,
   IPublishMessage,
-  ISignupData,
 } from "./utils";
