+++ packages/cli/ts/sdk/index.ts
@@ -4,36 +4,9 @@ import { genMaciPubKey } from "../commands/genPubKey";
 import { mergeSignups } from "../commands/mergeSignups";
 import { getPoll } from "../commands/poll";
 import { publish, publishBatch } from "../commands/publish";
-import {
-  signup,
-  isRegisteredUser,
-  getGatekeeperTrait,
-  getSemaphoreGatekeeperData,
-  getZupassGatekeeperData,
-  getEASGatekeeperData,
-  getHatsSingleGatekeeperData,
-  getMerkleProofGatekeeperData,
-} from "../commands/signup";
 import { verify } from "../commands/verify";
 
-export {
-  genKeyPair,
-  genMaciPubKey,
-  publish,
-  publishBatch,
-  signup,
-  isRegisteredUser,
-  verify,
-  getPoll,
-  extractVkToFile,
-  mergeSignups,
-  getGatekeeperTrait,
-  getSemaphoreGatekeeperData,
-  getZupassGatekeeperData,
-  getEASGatekeeperData,
-  getHatsSingleGatekeeperData,
-  getMerkleProofGatekeeperData,
-};
+export { genKeyPair, genMaciPubKey, publish, publishBatch, verify, getPoll, extractVkToFile, mergeSignups };
 
 export {
   linkPoseidonLibraries,
@@ -48,24 +21,14 @@ export * from "maci-contracts/typechain-types";
 
 export { VerifyingKey, PubKey, type IVkObjectParams } from "maci-domainobjs";
 
-export { GatekeeperTrait } from "../utils/interfaces";
-
 export type {
   TallyData,
   PublishArgs,
-  SignupArgs,
-  ISignupData,
   VerifyArgs,
   IGetPollArgs,
   IGetPollData,
-  IRegisteredUserArgs,
   IPublishBatchArgs,
   IGenKeypairArgs,
   IPublishBatchData,
   IPublishMessage,
-  IGetGatekeeperDataArgs,
-  ISemaphoreGatekeeperData,
-  IZupassGatekeeperData,
-  IHatsGatekeeperData,
-  IMerkleProofGatekeeperData,
 } from "../utils";
