+++ packages/cli/ts/utils/interfaces.ts
@@ -1,9 +1,6 @@
-import { MACI } from "maci-contracts/typechain-types";
-import { PubKey } from "maci-domainobjs";
-
 import type { Provider, Signer } from "ethers";
 import type { CircuitInputs } from "maci-core";
-import type { IMessageContractParams } from "maci-domainobjs";
+import type { PubKey, IMessageContractParams } from "maci-domainobjs";
 import type { Poll, SnarkProof } from "maci-sdk";
 import type { Groth16Proof, PublicSignals } from "snarkjs";
 
@@ -978,86 +975,6 @@ export interface SetVerifyingKeysArgs {
 }
 
 /**
- * Interface for the arguments to the signup command
- */
-export interface SignupArgs {
-  /**
-   * The public key of the user
-   */
-  maciPubKey: string;
-
-  /**
-   * A signer object
-   */
-  signer: Signer;
-
-  /**
-   * The address of the MACI contract
-   */
-  maciAddress: string;
-
-  /**
-   * The signup gateway data
-   */
-  sgDataArg?: string;
-
-  /**
-   * Whether to log the output
-   */
-  quiet?: boolean;
-}
-
-/**
- * Interface for the return data to the signup command
- */
-export interface ISignupData {
-  /**
-   * The state index of the user
-   */
-  stateIndex: string;
-
-  /**
-   * The voice credits of the user
-   */
-  voiceCredits: number;
-
-  /**
-   * The signup transaction hash
-   */
-  hash: string;
-}
-
-/**
- * Interface for the arguments to the register check command
- */
-export interface IRegisteredUserArgs {
-  /**
-   * A signer object
-   */
-  signer: Signer;
-
-  /**
-   * The public key of the user
-   */
-  maciPubKey: string;
-
-  /**
-   * The address of the MACI contract
-   */
-  maciAddress: string;
-
-  /**
-   * Start block for event parsing
-   */
-  startBlock?: number;
-
-  /**
-   * Whether to log the output
-   */
-  quiet?: boolean;
-}
-
-/**
  * Interface for the arguments to the get poll command
  */
 export interface IGetPollArgs {
@@ -1292,153 +1209,4 @@ export interface IParsePollJoinEventsArgs {
    * The public key
    */
   pollPublicKey: PubKey;
-}
-
-/**
- * Interface for the arguments to the parseSignupEvents function
- */
-export interface IParseSignupEventsArgs {
-  /**
-   * The MACI contract
-   */
-  maciContract: MACI;
-
-  /**
-   * The start block
-   */
-  startBlock: number;
-
-  /**
-   * The current block
-   */
-  currentBlock: number;
-
-  /**
-   * The public key
-   */
-  publicKey: PubKey;
-}
-
-/**
- * Interface for the arguments to the getGatekeeperTrait command
- */
-export interface IGetGatekeeperTraitArgs {
-  /**
-   * The address of the MACI contract
-   */
-  maciAddress: string;
-
-  /**
-   * A signer object
-   */
-  signer: Signer;
-}
-
-/**
- * Enum for the gatekeeper type
- */
-export enum GatekeeperTrait {
-  EAS = "EAS",
-  FreeForAll = "FreeForAll",
-  GitcoinPassport = "GitcoinPassport",
-  Hats = "Hats",
-  Semaphore = "Semaphore",
-  Token = "Token",
-  Zupass = "Zupass",
-  MerkleProof = "MerkleProof",
-}
-
-/**
- * Interface for the arguments to the get a gatekeeper's data command
- */
-export interface IGetGatekeeperDataArgs {
-  /**
-   * The address of the MACI contract
-   */
-  maciAddress: string;
-
-  /**
-   * A signer object
-   */
-  signer: Signer;
-}
-
-/**
- * Interface for the semaphore gatekeeper data
- */
-export interface ISemaphoreGatekeeperData {
-  /**
-   * The address of the semaphore gatekeeper
-   */
-  address: string;
-
-  /**
-   * The group ID
-   */
-  groupId: string;
-}
-
-/**
- * Interface for the zupass gatekeeper data
- */
-export interface IZupassGatekeeperData {
-  /**
-   * The event ID
-   */
-  eventId: string;
-
-  /**
-   * The first signer
-   */
-  signer1: string;
-
-  /**
-   * The second signer
-   */
-  signer2: string;
-}
-
-/**
- * Interface for the EAS gatekeeper data
- */
-export interface IEASGatekeeperData {
-  /**
-   * The EAS
-   */
-  eas: string;
-
-  /**
-   * The schema
-   */
-  schema: string;
-
-  /**
-   * The attester
-   */
-  attester: string;
-}
-
-/**
- * Interface for the Hats gatekeeper data
- */
-export interface IHatsGatekeeperData {
-  /**
-   * The criterion hat(s)
-   */
-  criterionHat: string[];
-
-  /**
-   * The hats contract
-   */
-  hatsContract: string;
-}
-
-/**
- * Interface for the MerkleProof gatekeeper data
- */
-export interface IMerkleProofGatekeeperData {
-  /**
-   * The merkle tree root
-   */
-  root: string;
 }
