+++ packages/sdk/ts/maci/index.ts
@@ -1,0 +1,18 @@
+export {
+  getGatekeeperTrait,
+  getSemaphoreGatekeeperData,
+  getZupassGatekeeperData,
+  getEASGatekeeperData,
+  getHatsSingleGatekeeperData,
+  getMerkleProofGatekeeperData,
+} from "./gatekeeper";
+export { EGatekeeperTrait } from "./types";
+export type {
+  IGetGatekeeperTraitArgs,
+  IGetGatekeeperDataArgs,
+  ISemaphoreGatekeeperData,
+  IZupassGatekeeperData,
+  IEASGatekeeperData,
+  IHatsGatekeeperData,
+  IMerkleProofGatekeeperData,
+} from "./types";
